Move Restore out of AnsiDB and into SQL specifically
Added not implemented exception for remaining ansi databases
