+++ grate/Migration/PostgreSqlDatabase.cs
@@ -1,4 +1,5 @@
 ﻿using System.Data.Common;
+using System.Threading.Tasks;
 using grate.Infrastructure;
 using Microsoft.Extensions.Logging;
 using Npgsql;
@@ -14,5 +15,10 @@ namespace grate.Migration
         public override bool SupportsDdlTransactions => true;
         public override bool SupportsSchemas => true;
         protected override DbConnection GetSqlConnection(string? connectionString) => new NpgsqlConnection(connectionString);
+
+        public override Task RestoreDatabase(string restoreFromPath)
+        {
+            throw new System.NotImplementedException("Restoring a database from file is not currently supported for  Postgresql.");
+        }
     }
 }
