+++ grate/Migration/SqlServerDatabase.cs
@@ -1,4 +1,7 @@
-﻿using System.Data.Common;
+﻿using System;
+using System.Data.Common;
+using System.Threading.Tasks;
+using System.Transactions;
 using grate.Infrastructure;
 using Microsoft.Data.SqlClient;
 using Microsoft.Extensions.Logging;
@@ -14,5 +17,42 @@ namespace grate.Migration
         public override bool SupportsDdlTransactions => true;
         public override bool SupportsSchemas => true;
         protected override DbConnection GetSqlConnection(string? connectionString) => new SqlConnection(connectionString);
+
+        public override async Task RestoreDatabase(string restoreFromPath)
+        {
+            try
+            {
+                await OpenAdminConnection();
+                Logger.LogInformation("Restoring {dbName} database on {server} server from path {path}.", DatabaseName, ServerName, restoreFromPath);
+                using var s = new TransactionScope(TransactionScopeOption.Suppress, TransactionScopeAsyncFlowOption.Enabled);
+                var cmd = AdminConnection.CreateCommand();
+                cmd.CommandText =
+                        $@"USE master
+                        ALTER DATABASE [{DatabaseName}] SET SINGLE_USER WITH ROLLBACK IMMEDIATE;
+                        RESTORE DATABASE [{DatabaseName}]
+                        FROM DISK = N'{restoreFromPath}'
+                        WITH NOUNLOAD
+                        , STATS = 10
+                        , RECOVERY
+                        , REPLACE;
+
+                        ALTER DATABASE [{DatabaseName}] SET MULTI_USER;";
+                await cmd.ExecuteNonQueryAsync();
+                s.Complete();
+            }
+            catch (Exception ex)
+            {
+                Logger.LogDebug(ex, "Got error: " + ex.Message);
+                throw;
+            }
+            finally
+            {
+                await CloseAdminConnection();
+            }
+
+            await WaitUntilDatabaseIsReady();
+
+            Logger.LogInformation("Database {dbName} successfully restored from path {path}.", DatabaseName, restoreFromPath);
+        }
     }
 }
